profile api

+++ models/pub/profile.js
@@ -13,17 +13,17 @@ module.exports.init = async (req, res) => {
 async function getProfileData(profileID) {
 	let raw = await db.select(`
 		SELECT
-			COUNT(*) as games,
-			COUNT(*) filter (WHERE tp_playergame.winner = true) as wins,
-			COUNT(*) filter (WHERE tp_playergame.winner = false) as losses,
+			COUNT(*)::REAL as games,
+			COUNT(*) filter (WHERE tp_playergame.winner = true)::REAL as wins,
+			COUNT(*) filter (WHERE tp_playergame.winner = false)::REAL as losses,
 
 			ROUND(COUNT(*) FILTER (WHERE tp_playergame.winner = true) * 100.0 / COUNT(*), 2)::REAL AS winrate,
 
-			SUM(cap_team_for) as CF,
-			SUM(cap_team_against) as CA,
-			SUM(cap_team_for - cap_team_against) as CD,
+			SUM(cap_team_for)::REAL as CF,
+			SUM(cap_team_against)::REAL as CA,
+			SUM(cap_team_for - cap_team_against)::REAL as CD,
 			TO_CHAR(SUM(duration) * interval '1 sec', 'hh24:mi:ss') as timeplayed,
-			ROUND(p.openskill::decimal, 2) as openskill
+			ROUND(p.openskill::decimal, 2)::REAL as openskill
 
 		FROM tp_playergame
 		LEFT JOIN tp_player as p ON p.id = tp_playergame.playerid
